Build page moved to dialog
-Build page is not a dialog, similar to the run view. This increases the amount of space that the graph can take up on mobile
-Added prompt to save progress before closing build (only works when you click the close icon, not refreshing or closing the tab)
