+++ packages/ui/src/components/containers/index.ts
@@ -1,5 +1,4 @@
 export * from './BuildBottomContainer/BuildBottomContainer';
-export * from './BuildInfoContainer/BuildInfoContainer';
 export * from './DialogActionsContainer/DialogActionsContainer';
 export * from './ListTitleContainer/ListTitleContainer';
 export * from './TitleContainer/TitleContainer';
