+++ packages/ui/src/components/containers/types.d.ts
@@ -50,21 +50,3 @@ export interface BuildBottomContainerProps {
     routine: Routine | null;
     runState: BuildRunState
 }
-
-export interface BuildInfoContainerProps {
-    canEdit: boolean;
-    handleAddLink: () => void;
-    handleLanguageUpdate: (language: string) => void;
-    handleNodeDelete: (nodeId: string) => void;
-    handleRoutineAction: (action: BaseObjectAction) => void;
-    handleRoutineUpdate: (changedRoutine: Routine) => void;
-    handleStartEdit: () => void;
-    handleTitleUpdate: (newTitle: string) => void;
-    isEditing: boolean;
-    language: string;
-    loading: boolean;
-    nodesOffGraph: Node[];
-    routine: Routine | null;
-    session: Session;
-    status: BuildStatusObject;
-}
