+++ packages/ui/src/components/graphs/NodeGraph/NodeGraph/NodeGraph.tsx
@@ -415,12 +415,10 @@ export const NodeGraph = ({
             cursor: isShiftKeyPressed ? 'nesw-resize' : 'move',
             minWidth: '100%',
             // Graph fills remaining space that is not taken up by other elements. 
-            // On small screens, these are: navbar (64px), routine title (64px), other top build icons (48px),
-            // build bottom (48px), and bottom navbar (56px). This makes the size: 
-            // 100vh - (64 + 64 + 48 + 48 + 56) = calc(100vh - 280px).
-            // On larger screens, these are: navbar (80px), other top build icons (48px), and build bottom (48px).
-            // This makes the size 100vh - (80 + 48 + 48) = calc(100vh - 176px).
-            height: { xs: 'calc(100vh - 280px)', md: 'calc(100vh - 240px)', lg: 'calc(100vh - 176px)' },
+            // These are: routine title (64px), other top build icons (48px),
+            // build bottom (48px). This makes the size: 
+            // 100vh - (64 + 48 + 48) = calc(100vh - 216px).
+            height: 'calc(100vh - 160px)',
             overflowX: 'auto',
             overflowY: 'auto',
             margin: 0,
