+++ packages/ui/src/pages/index.ts
@@ -3,7 +3,6 @@ export * from './search';
 export * from './view';
 export * from './wrapper';
 export * from './NotFoundPage';
-export * from './BuildPage/BuildPage';
 export * from './SettingsPage/SettingsPage';
 export * from './StartPage/StartPage';
 export * from './WelcomePage/WelcomePage';
