+++ packages/ui/src/pages/types.d.ts
@@ -1,9 +1,5 @@
 import { Session } from 'types';
 
-export interface BuildPageProps {
-    session: Session;
-}
-
 export interface SettingsPageProps {
     session: Session;
 }
